+++ drivers/wireless/src/rtl8xxxu/rtl8822b.rs
@@ -80,7 +80,7 @@ pub const N_RF_B_ROWS: usize = 245;
 
 /// Stage 0/1 register init table.
 pub const INIT_TABLE: &[(u16, u8)] = &[
-    (REG_APS_FSMCO as u16 + 1, 0x08),
+    (REG_APS_FSMCO + 1, 0x08),
     (
         REG_CR,
         (CR_HCI_TXDMA_ENABLE
@@ -359,7 +359,7 @@ pub fn init_rf<W: FnMut(u8, u8, u32)>(mut write_rfreg: W) -> usize {
 
 /// Build USB control-transfer setup for APS_FSMCO MAC enable.
 pub fn aps_fsmco_mac_enable_setup() -> UsbControlSetup {
-    UsbControlSetup::write(REG_APS_FSMCO as u16 + 1, 1)
+    UsbControlSetup::write(REG_APS_FSMCO + 1, 1)
 }
 
 /// Build bulk-OUT TX frame with a 40-byte descriptor prefix.
